requirements.txt
fastapi
uvicorn
python-multipart
pillow 
opencv-python
matplotlib
pandas